Implements login view with new logo and styling
Adds a new login view with username and password fields and a login button.
Updates the logo image and adds styling to the login button.
The old login-view.fxml and DiabedibileLogo.png files have been deleted.
Adds a new home-diabetic view, that contains the layout of the home page for the diabetic user